backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
boto3>=1.35.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
requests>=2.32.3
mcp>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.1
numpy>=1.26.0
openai>=1.0.0
